Reduce to support single statement source for now
Compiles, but still needs further refactoring because previously the
aggregration call was passed in by engine.go, which occurred inside
processAggregate(), by the MR execute call. This seems to indicate that
doing so much within Begin() is wrong. This also is indicated by the
amount of own-variable setting going on within that function. That
should be set by a function creating the Mappers.